defined product property
this is the step i'm most iffy about... i'm not exactly sure what we accomplished here and the tutorial doesn't explain it very clearly. why are components set up as classes and how exactly does this syntax "define" a "property" (whatever we mean by that)? i don't have much if any experience with javascript and a bunch of the stuff in this part looks beyond foreign to me... some insight or slowing down a bit here would be nice